citizen_assurance_idv/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pytesseract
opencv-python
face-recognition
